frozen-requirements.txt
acme==1.21.0
amqp==5.2.0
ansible-base==2.10.8
apache-libcloud==3.2.0
APScheduler==3.10.4
argcomplete==1.8.1
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
billiard==4.2.1
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
Brlapi==0.8.3
celery==5.4.0
certbot==1.21.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==23.4.4
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
ec2-hibinit-agent==1.0.0
Flask==3.0.3
gyp==0.1
hibagent==1.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
install==1.3.5
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
josepy==1.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kombu==5.4.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.5
more-itertools==8.10.0
mysql-connector-python==8.4.0
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
parsedatetime==2.6
pexpect==4.8.0
Pillow==9.0.1
prompt_toolkit==3.0.48
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.6
reportlab==3.6.8
requests==2.25.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
rq==1.16.2
s3transfer==0.10.1
SecretStorage==3.3.1
selinux==3.3
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sos==4.5.6
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
tzdata==2024.2
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.3
xdg==5
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
